Move internal http package to instrumentation
