+++ instrumentation/sources/labstack/echo.v4/middleware.go
@@ -3,7 +3,7 @@ package echo
 import (
 	"errors"
 
-	"github.com/AikidoSec/firewall-go/internal/http"
+	"github.com/AikidoSec/firewall-go/instrumentation/http"
 	"github.com/AikidoSec/firewall-go/internal/request"
 	"github.com/AikidoSec/firewall-go/zen"
 	"github.com/labstack/echo/v4"
