+++ instrumentation/sources/net/http/middleware.go
@@ -6,7 +6,7 @@ import (
 	"strings"
 	"sync/atomic"
 
-	zenhttp "github.com/AikidoSec/firewall-go/internal/http"
+	zenhttp "github.com/AikidoSec/firewall-go/instrumentation/http"
 	"github.com/AikidoSec/firewall-go/internal/log"
 	"github.com/AikidoSec/firewall-go/internal/request"
 	"github.com/AikidoSec/firewall-go/zen"
